requirements.txt
tkmacosx == 1.0.5 #available only for macOs 
Pillow == 10.0.0